Changed id types from Long to Integers since it's overkill for an application this size.

+++ src/main/java/com/dj/model/Company.java
@@ -5,9 +5,9 @@ package com.dj.model;
  */
 public interface Company {
 	
-	public void setId(long id);
+	public void setId(int id);
 	
-	public long getId();
+	public int getId();
 	
 	public void setName(String name);
