+++ src/main/java/com/dj/repository/CompanyCountryRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * Created by DJ on 11/23/16.
  */
 @Repository
-public interface CompanyCountryRepository extends JpaRepository<CompanyCountry,Long> {
+public interface CompanyCountryRepository extends JpaRepository<CompanyCountry,Integer> {
 	
 }
