+++ src/main/java/com/dj/repository/SystemRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * Created by DJ on 11/23/16.
  */
 @Repository
-public interface SystemRepository extends JpaRepository<System,Long> {
+public interface SystemRepository extends JpaRepository<System,Integer> {
 	
 }
